step1.2 mostly done, somehow delete returns 200 detected as error though:
client-side this is causing the "error" callback to activate, rather than the "success" callback on

res.status(200).send("Movie saved successfully");

==> The client-side will accept a string result for an error-status code, but not for a 200 success-status code.
For a 200-status response, the result-value must be encoded in JSON format.

Typically, a 200-status result-value is a JSON representation of a model or collection. In a few cases, such as a delete-handler, the result can be a JSON-encoded string, such as:

{"responseText": "movie successfully deleted"}